demo/requirements.txt
torch
torchvision
torchmetrics
zenml
click
pyyaml
torchvision
lightning